@@ -15,27 +15,28 @@ __license__ = 'Apache License 2.0'
 
 GITHUB_URL = 'https://github.com/WZBSocialScienceCenter/tmtoolkit'
 
-DEPS_BASE = ['numpy>=1.20.0,<2', 'scipy>=1.6.0,<1.7', 'pandas>=1.2.0,<1.3', 'xlrd>=2.0.0,<2.1', 'openpyxl>=3.0.0,<3.1',
+DEPS_BASE = ['numpy>=1.22.0', 'scipy>=1.7.0', 'globre>=0.1.5',
+             'pandas>=1.3.0', 'xlrd>=2.0.0', 'openpyxl>=3.0.0',
              'matplotlib>=3.5.0']
 
 DEPS_EXTRA = {
-    'textproc': ['spacy>=2.3.0,<2.4', 'globre>=0.1.5,<0.2'],
-    'textproc-extra': ['PyICU>=2.7,<2.8', 'nltk>=3.6.0,<3.7'],
-    'wordclouds': ['wordcloud>=1.8.0,<1.9', 'Pillow>=8.2.0,<8.3'],
-    'lda': ['lda>=2.0,<2.1'],
-    'sklearn': ['scikit-learn>=0.24,<0.25'],
-    'gensim': ['gensim>=4.0.0,<4.1'],
-    'topic_modeling_eval_extra': ['gmpy2>=2.0.0,<3'],
-    'test': ['pytest>=6.0.0,<7', 'hypothesis>=6.13.0<6.14', 'decorator>=5.0.0,<5.1'],
-    'doc': ['Sphinx>=4.0.0', 'sphinx-rtd-theme>=0.5.0', 'nbsphinx>=0.7.0'],
-    'dev': ['coverage>=5.2', 'coverage-badge>=1.0.0', 'pytest-cov>=2.10.0', 'twine>=3.2.0',
-            'ipython>=7.16.0', 'jupyter>=1.0.0', 'notebook>=6.0.0', 'tox>=3.18.0'],
+    'textproc': ['spacy>=3.2.0', 'bidict>=0.21.0', 'loky>=3.0.0'],
+    'textproc_extra': ['PyICU>=2.8', 'nltk>=3.6.0'],
+    'wordclouds': ['wordcloud>=1.8.0,<1.9', 'Pillow>=9.0.0'],
+    'lda': ['lda>=2.0'],
+    'sklearn': ['scikit-learn>=1.0.0'],
+    'gensim': ['gensim>=4.1.0'],
+    'topic_modeling_eval_extra': ['gmpy2>=2.1.0'],
+    'test': ['pytest>=6.2.0', 'hypothesis>=6.35.0'],
+    'doc': ['Sphinx>=4.3.0', 'sphinx-rtd-theme>=1.0.0', 'nbsphinx>=0.8.0'],
+    'dev': ['coverage>=6.2', 'coverage-badge>=1.1.0', 'pytest-cov>=3.0.0', 'twine>=3.7.0',
+            'ipython>=7.31.0', 'jupyter>=1.0.0', 'notebook>=6.4.0', 'tox>=3.24.0'],
 }
 
 # DEPS_EXTRA['minimal'] = DEPS_BASE   # doesn't work with extras_require and pip currently
 # see https://github.com/pypa/setuptools/issues/1139
 
-DEPS_EXTRA['recommended'] = DEPS_EXTRA['textproc']
+DEPS_EXTRA['recommended'] = DEPS_EXTRA['textproc'] + DEPS_EXTRA['wordclouds']
 DEPS_EXTRA['all'] = []
 for k, deps in DEPS_EXTRA.items():
     if k not in {'recommended', 'all'}:
@@ -73,9 +74,9 @@ setup(
         'Operating System :: OS Independent',
         'Programming Language :: Python',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.6',
-        'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
+        'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
 
         'Topic :: Scientific/Engineering :: Information Analysis',
         'Topic :: Software Development :: Libraries :: Python Modules',
@@ -86,7 +87,7 @@ setup(
 
     packages=find_packages(exclude=['tests', 'examples']),
     include_package_data=True,
-    python_requires='>=3.6',
+    python_requires='>=3.8',
     install_requires=DEPS_BASE,
     extras_require=DEPS_EXTRA
 )
